test: extract systemjs-exec test into single file
